requirements.txt
django
torch
torchvision
torchaudio
pillow
numpy